Updated guidance on systemofRecords property to account for how these URLs will be provided.
SORNs are published in the Federal Register, and may be published in batches.  When linking to these kinds of situations, the user may be confused because they won't know which SORN to look for. Providing the identifier or the title of the SORN as well as the URL will be clearer.

Schema should continue to validate this as a string rather than a URL.