Update CONTRIBUTING.md
No where on the github repo could I find a link to the site. I had to look through the pull requests to find out how to see the content. I added a link on the contribute page that provides at least a little information on how to get to the site.