Update index.md
As suggested, the "larger files" clause has been removed.  json-csv.com has been in operation since May 2012 so it has likely dealt with a lot more users over this time to address a lot of issues.  Converting JSON to CSV is subjective given that it transforms a 3D data structure into 2D.  I believe that providing a few alternatives will be of value for the purpose of this document.