Pedantic Proposal - Mathmatical Symbols
I would propose changing the Cardinality symbol from <parens> to <braces>, i.e. '{' or '}'.

This is because Cardinality is part of set theory: http://en.wikipedia.org/wiki/Cardinality

Where as the "( )" can be confused with inclusive|exclusive ranges:
http://stackoverflow.com/questions/4396290/what-does-this-square-bracket-and-parenthesis-bracket-notation-mean-first1-last

I will change all of them in a fork if necessary but wasn't sure if this was auto-generated documentation.