reorganizing around event type
This is just reordering to consolidate sections around the event type;
I haven’t made any other changes to the text in this commit. I was
recently reviewing this and it was hard to get a sense of what each
event and what they involved in the current format. Planning to add
some content later…