Making contribution instructions more precise
Here are a few suggestions for this page, since I read it out of curiosity and noticed a couple bits of phrasing you could adjust. I saw that this contribution guide suggests clicking a button that says "Improve this content", while the button on the top of the page says "Edit this Page", so I updated that text.

I also added a little more detail to the instructions, since outside contributors will get prompted to fork the repository, and the form on GitHub's website says "Propose file change" instead of "Submit".

I changed "very easy" to "only takes a few steps", and I removed "simply" - since readers may find that this process isn't easy or simple for them, it can be more empathetic to skip calling it easy and simple. :)

Hope this helps!